Add alternate layout options

Adds a layout that will be triggered based upon the size of the viewport. This is typically used to add a mobile view (or a desktop view if your app is mobile first.)

add_alternate_layout(
  layout,
  alternate_layout,
  width_bounds = NULL,
  container_height = NULL
)

Arguments

layout

Main (or default) layout object

alternate_layout

A gridlayout object or layout-as-markdown-table defining layout.

width_bounds

A named vector with at least one element of name min, or max. These two values are used to set when the layout occurs. For instance, c(max = 600) means the layout will occur up until the page is wider than 600px.

container_height

How tall the layout should be. If left as NULL this will simply inherit the main layout's height. Often this should be set to "auto" to allow mobile user's to scroll. Beware though, you will want to absolutely size any plot outputs in these cases.