Get elements and positions out of a grid layout

Get elements and positions out of a grid layout

get_elements(layout)

Arguments

layout
Object of class "gridlayout".

Value

List of all unique elements in the layout and their id, {start,end}_row, and {start,end}_col. Positions are indexed starting at 1

See also

Examples

grid_obj <- md_to_gridlayout(
  layout_table = "
        |      |120px   |1fr    |1fr    |
        |------|--------|-------|-------|
        |100px |header  |header |header |
        |1fr   |sidebar |plot_a |plot_a |
        |1fr   |sidebar |plot_b |plot_c |"
 )

 get_elements(grid_obj)

#> [[1]]
#> [[1]]$id
#> [1] "header"
#> 
#> [[1]]$start_row
#> [1] 1
#> 
#> [[1]]$end_row
#> [1] 1
#> 
#> [[1]]$start_col
#> [1] 1
#> 
#> [[1]]$end_col
#> [1] 3
#> 
#> 
#> [[2]]
#> [[2]]$id
#> [1] "sidebar"
#> 
#> [[2]]$start_row
#> [1] 2
#> 
#> [[2]]$end_row
#> [1] 3
#> 
#> [[2]]$start_col
#> [1] 1
#> 
#> [[2]]$end_col
#> [1] 1
#> 
#> 
#> [[3]]
#> [[3]]$id
#> [1] "plot_a"
#> 
#> [[3]]$start_row
#> [1] 2
#> 
#> [[3]]$end_row
#> [1] 2
#> 
#> [[3]]$start_col
#> [1] 2
#> 
#> [[3]]$end_col
#> [1] 3
#> 
#> 
#> [[4]]
#> [[4]]$id
#> [1] "plot_b"
#> 
#> [[4]]$start_row
#> [1] 3
#> 
#> [[4]]$end_row
#> [1] 3
#> 
#> [[4]]$start_col
#> [1] 2
#> 
#> [[4]]$end_col
#> [1] 2
#> 
#> 
#> [[5]]
#> [[5]]$id
#> [1] "plot_c"
#> 
#> [[5]]$start_row
#> [1] 3
#> 
#> [[5]]$end_row
#> [1] 3
#> 
#> [[5]]$start_col
#> [1] 3
#> 
#> [[5]]$end_col
#> [1] 3
#> 
#> 
#> attr(,"row_sizes")
#> [1] "100px" "1fr"   "1fr"  
#> attr(,"col_sizes")
#> [1] "120px" "1fr"   "1fr"  
#> attr(,"gap")
#> [1] "1rem"
#> attr(,"container_height")
#> [1] "viewport"
#> attr(,"alternates")
#> attr(,"alternates")[[1]]
#> attr(,"alternates")[[1]]$layout
#> gridlayout with 5 rows, 1 columns, and gap size of 1rem. Total height of auto.
#>       1fr    
#> 85px  header 
#> 350px sidebar
#> 350px plot_a 
#> 350px plot_b 
#> 350px plot_c 
#> attr(,"alternates")[[1]]$width_bounds
#>     max 
#> "600px" 
#> 
#> 
#> attr(,"auto_mobile_layout")
#> [1] TRUE