Grid panel with vertically stacked elements

Contain a series of ui elements with each vertically stacked on top of eachother. Internally uses css flexbox to align items.

grid_panel_stack(
  area,
  ...,
  item_alignment = "top",
  title = NULL,
  collapsible = TRUE,
  scrollable = FALSE,
  item_gap = "10px"
)

Arguments

area

Name of grid area, should match an area defined in the layout section of the wrapping grid_page() or grid_container().

...

Elements to include within the panel

item_alignment

How should the items within the panel be aligned. Defaults to stacking downward from the top. Options include "top", "centered", "bottom", or "spread" (items are distributed evenly among vertical space).

title

Character string to go across top of panel with label. If left blank the card contents will take up entire space.

collapsible

Should the card be able to be collapsed (TRUE or FALSE)? Gridlayout will only show collapser if the layout allows it (panel is entirely positioned within "auto" sized rows, and has a title). Setting this to FALSE will mean collapsibility of the panel will never be enabled, regardless of layout.

scrollable

Should scroll-bars be added so content that is larger than the panel can be seen?

item_gap

How much space should there be between consecutive items?

Value

Elements from ... wrapped in a shiny::div() with styles for vertical stacking applied.

See also

Examples

# Simply a wrapper for shiny::div(...)
grid_panel_stack(
  area="header",
  shiny::h2("R"),
  shiny::actionButton("myButton", "Click Me")
)
#> <div class="grid_panel vertical_stack" style="grid-area:header;">
#>   <div style="--item-gap:10px;justify-content:flex-start;" class="panel-content">
#>     <h2>R</h2>
#>     <button id="myButton" type="button" class="btn btn-default action-button">Click Me</button>
#>   </div>
#> </div>