Construct a gridlayout object

Builds the gridlayout s3 class that holds information needed to draw a given layout.

new_gridlayout(
  layout_def = list(),
  col_sizes = NULL,
  row_sizes = NULL,
  gap = NULL,
  container_height = NULL,
  alternate_layouts = "auto"
)

Arguments

layout_def
Either a list of elements with the id, start_row, end_row, start_col, and end_col format, or a markdown table defining a layout.

col_sizes
A character vector of valid css sizes for the width of each column in your grid as given by layout_mat. If a single value is passed, it will be repeated for all columns.

row_sizes
Same as col_sizes, but for row heights.

gap
Valid css sizing for gap to be left between each element in your grid. Defaults to "1rem". This is a relative unit that scales with the base text size of a page. E.g. setting font-size: 16px on the body element of a page means 1rem = 16px;

container_height
Valid css unit determining how tall the containing element should be for this layout. Defaults to "viewport" (full page height: equivalent to the CSS value of 100vh) if any relative units (e.g. fr or auto) are included in row sizes and auto otherwise. Values such as "auto" will let the page grow to as large as it needs to be to fit all content. This should most likely be avoided when using row heights in relative units.

alternate_layouts
A list of layouts to be used for different viewport widths. This is enables your app to adapt to different screensizes such as a phone or an ultra-wide monitor. Each entry in this list must contain a layout: or valid layout declaration (see Declaring your layout), and width_bounds: or a list with at least one of a min and max value for when your page appears. See add_alternate_layout() for more details. If no alternate layouts are given a single-column layout will be automatically applied for mobile screens (viewports less than 600px wide). Set to NULL to avoid this.

Value

Object of class "gridlayout"

Declaring your layout

There are three current ways to declare layouts (aka inputs to layout_def).

Markdown tables The first is the easiest: a markdown table wrapped in a

character string. In this format you define a grid using the table and then place your grid "elements" within that grid using their grid id. So for a 2x2 layout with a header along the top and two plots side-by-side the layout would look as follows: {r} new_gridlayout( "| header | header | | plota | plotb |" )

Element lists The second method is to supply a list of elements by

providing the following information for each:

id: Identifying id of the element (e.g. "header) - start_row/end_row: The (1-indexed) start and end row of your element's span - start_col/end_col: The start and end column for your element's span

This is a bit more verbose but allows for greater control. Here you can have overlapping elements etc.. Most of the time you will just want to use the markdown syntax

The same layout as declared above can be accomplished with the following:

start_col = 1, end_col = 2), list(id = "plot_a", start_row = 2, end_row = 2,
start_col = 1, end_col = 1), list(id = "plot_b", start_row = 2, end_row = 2,
start_col = 2, end_col = 2) ) ) ```

### An existing `gridlayout`

The last way is to pass an existing `gridlayout` object in. This allows you to
do things like modify the grid sizing or container sizes of an existing
layout.

Examples

# Assemble list of elements along with their positions
elements_list <- list(
  list(id = "header", start_row = 1, end_row = 1,
       start_col = 1, end_col = 2),
  list(id = "plot",   start_row = 2, end_row = 2,
       start_col = 1, end_col = 1),
  list(id = "table",  start_row = 2, end_row = 2,
       start_col = 2, end_col = 2),
  list(id = "footer", start_row = 3, end_row = 3,
       start_col = 1, end_col = 2)
)

new_gridlayout(
  elements_list,
  col_sizes = c("1fr", "2fr"),
  row_sizes = c("100px", "1fr", "1fr")
)

#> gridlayout with 3 rows, 2 columns, and gap size of 1rem. Total height of viewport.
#>       1fr    2fr   
#> 100px header header
#> 1fr   plot   table 
#> 1fr   footer footer
#> 
#> Alternative layouts:
#> 
#> When width of page < 600px 
#> gridlayout with 4 rows, 1 columns, and gap size of 1rem. Total height of auto.
#>       1fr   
#> 300px header
#> 300px plot  
#> 300px table 
#> 300px footer

# Can also use a matrix for more visually intuitive laying out
new_gridlayout(
  layout_def = "
      | header | header |
      | plota  | plotb  |",
  col_sizes = c("1fr", "2fr"),
  row_sizes = c("100px", "1fr"),
  gap = "2rem"
)

#> gridlayout with 2 rows, 2 columns, and gap size of 2rem. Total height of viewport.
#>       1fr    2fr   
#> 100px header header
#> 1fr   plota  plotb 
#> 
#> Alternative layouts:
#> 
#> When width of page < 600px 
#> gridlayout with 3 rows, 1 columns, and gap size of 1rem. Total height of auto.
#>       1fr   
#> 300px header
#> 300px plota 
#> 300px plotb