Panel just containing text

Makes a grid_panel that contains just text that is vertically centered within the panel. Useful for titles (see title_panel()) or displaying text-based statistics.

text_panel(
  content,
  ...,
  h_align = "start",
  img_height = "55px",
  icon = NULL,
  wrapping_tag = "h2"
)

Arguments

content
Whatever you want the title to say. Typically just text but any tag or tag-list is possible. All will get wrapped in an h3 tag.

...
Arguments passed on to grid_panel

title

Character string to go across top of panel with label. If left blank the card contents will take up entire space.

use_card_styles

Should the element contained in panel be made to look like an enclosed card?

use_bslib_card_styles

Should the card styles from bslib be used instead of default gridlayout card styles? If this is set to TRUE it will override use_card_style.

collapsible

Should the card be able to be collapsed (TRUE or FALSE)? Gridlayout will only show collapser if the layout allows it (panel is entirely positioned within "auto" sized rows, and has a title). Setting this to FALSE will mean collapsibility of the panel will never be enabled, regardless of layout.

scrollable

Should scroll-bars be added so content that is larger than the panel can be seen?

v_align

Vertical and Horizontal alignment of content in panel. Options include "center", "start", "end", "stretch". These are a direct mapping to the the css-spec for justify-items (=h_align) and align-items (= v_align).

panel_id

ID of grid_panel div. Usually this is automatically set by grid_page() or grid_container().

panel_class

Class(es) of grid_panel div. Like panel_id this is typically automatically generated and should only be touched in advanced usage.

h_align
Vertical and Horizontal alignment of content in panel. Options include "center", "start", "end", "stretch". These are a direct mapping to the the css-spec for justify-items (=h_align) and align-items (= v_align).

img_height
If the passed icon is a path to an image, how tall should that image be rendered (preserves aspect ratio.)

icon
Optional icon/image for left of text. Supports image locations (those ending in .png, .jpg, .jpeg or .svg), ids of font-awesome icons (i.e. that works with fontawesome::fa(icon)], or fontawesome icons as returned by fontawesome::fa() (if customization of icon style is desired.)

wrapping_tag
What tag should the text be wrapped in. Takes either an htmltools tag function or the string of a tag accessible via htmltools::tags[[wrapping_tag]].

Examples

# Typically you'll just pass a character string to the function
title_panel("This is my header")

#> <div class="grid_panel gridlayout-card">
#>   <div style="display:grid;justify-content:start;align-content:center;" class="panel-content">
#>     <h2 class="text_panel">This is my header</h2>
#>   </div>
#> </div>

# Icons from `fontawesome` can be used:

# Either with just the id
text_panel("Here's my text", icon = "r-project")

#> Loading required namespace: fontawesome

#> <div class="grid_panel gridlayout-card">
#>   <div style="display:grid;justify-content:start;align-content:center;" class="panel-content">
#>     <h2 class="text_panel">
#>       <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 581 512" class="rfa" style="height:0.75em;fill:currentColor;position:relative;"><path d="M581 226.6C581 119.1 450.9 32 290.5 32S0 119.1 0 226.6C0 322.4 103.3 402 239.4 418.1V480h99.1v-61.5c24.3-2.7 47.6-7.4 69.4-13.9L448 480h112l-67.4-113.7c54.5-35.4 88.4-84.9 88.4-139.7zm-466.8 14.5c0-73.5 98.9-133 220.8-133s211.9 40.7 211.9 133c0 50.1-26.5 85-70.3 106.4-2.4-1.6-4.7-2.9-6.4-3.7-10.2-5.2-27.8-10.5-27.8-10.5s86.6-6.4 86.6-92.7-90.6-87.9-90.6-87.9h-199V361c-74.1-21.5-125.2-67.1-125.2-119.9zm225.1 38.3v-55.6c57.8 0 87.8-6.8 87.8 27.3 0 36.5-38.2 28.3-87.8 28.3zm-.9 72.5H365c10.8 0 18.9 11.7 24 19.2-16.1 1.9-33 2.8-50.6 2.9v-22.1z"/></svg>
#>       Here's my text
#>     </h2>
#>   </div>
#> </div>

# Or by directly passing the icon object if you want more customization
text_panel("Here's my text", icon = fontawesome::fa("r-project", fill = "steelblue"))

#> <div class="grid_panel gridlayout-card">
#>   <div style="display:grid;justify-content:start;align-content:center;" class="panel-content">
#>     <h2 class="text_panel">
#>       <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 581 512" class="rfa" style="height:0.75em;fill:steelblue;position:relative;"><path d="M581 226.6C581 119.1 450.9 32 290.5 32S0 119.1 0 226.6C0 322.4 103.3 402 239.4 418.1V480h99.1v-61.5c24.3-2.7 47.6-7.4 69.4-13.9L448 480h112l-67.4-113.7c54.5-35.4 88.4-84.9 88.4-139.7zm-466.8 14.5c0-73.5 98.9-133 220.8-133s211.9 40.7 211.9 133c0 50.1-26.5 85-70.3 106.4-2.4-1.6-4.7-2.9-6.4-3.7-10.2-5.2-27.8-10.5-27.8-10.5s86.6-6.4 86.6-92.7-90.6-87.9-90.6-87.9h-199V361c-74.1-21.5-125.2-67.1-125.2-119.9zm225.1 38.3v-55.6c57.8 0 87.8-6.8 87.8 27.3 0 36.5-38.2 28.3-87.8 28.3zm-.9 72.5H365c10.8 0 18.9 11.7 24 19.2-16.1 1.9-33 2.8-50.6 2.9v-22.1z"/></svg>
#>       Here's my text
#>     </h2>
#>   </div>
#> </div>

# You can also pass arbitrary image locations for the icon
text_panel("Here's my text", icon = "https://cran.r-project.org/Rlogo.svg")

#> <div class="grid_panel gridlayout-card">
#>   <div style="display:grid;justify-content:start;align-content:center;" class="panel-content">
#>     <h2 class="text_panel">
#>       <img src="https://cran.r-project.org/Rlogo.svg" height="55px"/>
#>       Here's my text
#>     </h2>
#>   </div>
#> </div>

# These images can have their size controlled
text_panel("Here's my text", icon = "https://cran.r-project.org/Rlogo.svg", img_height = "20px")

#> <div class="grid_panel gridlayout-card">
#>   <div style="display:grid;justify-content:start;align-content:center;" class="panel-content">
#>     <h2 class="text_panel">
#>       <img src="https://cran.r-project.org/Rlogo.svg" height="20px"/>
#>       Here's my text
#>     </h2>
#>   </div>
#> </div>

# You can adjust the horizontal alignment of your header with h_align
text_panel("I'm in the middle", h_align = "center")

#> <div class="grid_panel gridlayout-card">
#>   <div style="display:grid;justify-content:center;align-content:center;" class="panel-content">
#>     <h2 class="text_panel">I'm in the middle</h2>
#>   </div>
#> </div>