Make header panel

This is just a helper function that wraps your content in a vertically centered header (h2) tag via the text_panel() function. Also adds the apps content to the title metatag of your webpage so it shows up properly in tabs etc.. Modeled after shiny::titlePanel().

title_panel(
  area,
  title,
  windowTitle = title,
  h_align = "start",
  img_height = "55px",
  icon = NULL
)

Arguments

area

Name of grid area, should match an area defined in the layout section of the wrapping grid_page() or grid_container().

title

An application title to display.

windowTitle

The title that should be displayed by the browser window.

h_align

Vertical and Horizontal alignment of content in panel. Options include "center", "start", "end", "stretch". These are a direct mapping to the the css-spec for justify-items (=h_align) and align-items (= v_align).

img_height

If the passed icon is a path to an image, how tall should that image be rendered (preserves aspect ratio.)

icon

Optional icon/image for left of text. Supports image locations (those ending in .png, .jpg, .jpeg or .svg), ids of font-awesome icons (i.e. that works with fontawesome::fa(icon)], or fontawesome icons as returned by fontawesome::fa() (if customization of icon style is desired.)

See also

Examples

# Simple app title
title_panel(area = "header", "My App")
#> <div class="grid_panel gridlayout-card" style="grid-area:header;" data-gridlayout-area="header">
#>   <div style="display:grid;justify-content:start;align-content:center;" class="panel-content">
#>     <h2 class="text_panel">My App</h2>
#>   </div>
#> </div>

# Centered title
title_panel("header", "My App (in the center)", h_align = "center")
#> <div class="grid_panel gridlayout-card" style="grid-area:header;" data-gridlayout-area="header">
#>   <div style="display:grid;justify-content:center;align-content:center;" class="panel-content">
#>     <h2 class="text_panel">My App (in the center)</h2>
#>   </div>
#> </div>

# Can have a logo
title_panel("footer", "My App", icon = "https://cran.r-project.org/Rlogo.svg")
#> <div class="grid_panel gridlayout-card" style="grid-area:footer;" data-gridlayout-area="footer">
#>   <div style="display:grid;justify-content:start;align-content:center;" class="panel-content">
#>     <h2 class="text_panel">
#>       <img src="https://cran.r-project.org/Rlogo.svg" height="55px"/>
#>       My App
#>     </h2>
#>   </div>
#> </div>