Convert grid layout to css string

Convert grid layout to css string

to_css(layout, ...)

Arguments

layout
Object of class "gridlayout".

container
Id of the element for grid to be placed in. Default value to apply grid styling to the whole app (aka the body element) for whole page grids. If an simple name is passed it is assumed that it is an id: i.e. the css id selector # is apended to the name. If container contains css selector characters such as a dot, the selector will not be transformed into an id automatically. E.g. container = ".main-content".

is_card_styled
Should each section of the grid be given a card style to make it stand out? Options are "grid_panel", where only elements with "grid_panel" class will get card styling, "all" where all children of the grid container will get card styling (useful for RMarkdown or other situations where you don't control child rendering) or "none" for no card styling.

element_styles
A list of named property-value pairs ( E.g. element_styles = c("background" = "blue").) for additional styles to be added to each first child element of the grid container. E.g. under the css selector of grid_container > *. If is_card_styled = "all" then this has the same purpose as as additional_card_styles

selector_prefix
CSS prefix used to target grid elements. This will change if you're integrating grid with a system that you don't want to use ids (the "#" prefix) with because they are not available or are used for other reasons.

Value

Character string of css used to setup grid layout and place elements (referenced by id) into correct locations

Examples

grid_obj <- md_to_gridlayout(
  layout_table = "
    |      |120px   |1fr    |1fr    |
    |:-----|:-------|:------|:------|
    |100px |header  |header |header |
    |1fr   |sidebar |plot_a |plot_c |
    |1fr   |sidebar |plot_b |plot_b |"
)

cat(to_css(grid_obj))

#> 
#> 
#> body {
#>   display: grid;
#>   grid-template-rows: 100px 1fr 1fr;
#>   grid-template-columns: 120px 1fr 1fr;
#>   grid-gap: 1rem;
#>   padding: 1rem;
#>   height: 100vh;
#> }
#> 
#> #header {
#>   grid-column-start: 1;
#>   grid-column-end: 4;
#>   grid-row-start: 1;
#>   grid-row-end: 2;
#> }
#> #sidebar {
#>   grid-column-start: 1;
#>   grid-column-end: 2;
#>   grid-row-start: 2;
#>   grid-row-end: 4;
#> }
#> #plot_a {
#>   grid-column-start: 2;
#>   grid-column-end: 3;
#>   grid-row-start: 2;
#>   grid-row-end: 3;
#> }
#> #plot_b {
#>   grid-column-start: 2;
#>   grid-column-end: 4;
#>   grid-row-start: 3;
#>   grid-row-end: 4;
#> }
#> #plot_c {
#>   grid-column-start: 3;
#>   grid-column-end: 4;
#>   grid-row-start: 2;
#>   grid-row-end: 3;
#> }
#> 
#> 
#> 
#> .grid_panel {
#>   --card-padding: 0.8rem;
#>   box-sizing: border-box;
#>   overflow: hidden;
#>   display: grid;
#>   grid-template-areas: "title"
#>                        "content";
#>   /* When there's no title the column will dissapear */
#>   grid-template-rows: min-content 1fr;
#> }
#> 
#> .grid_panel .title-bar {
#>   grid-area: title;
#>   height: 100%;
#>   width: 100%;
#>   display: flex;
#>   justify-content: start;
#>   align-items: center;
#>   border-bottom: 1px solid rgba(0,0,0,0.125);
#>   padding: calc(var(--card-padding)/2) var(--card-padding);
#> }
#> .grid_panel .title-bar > h3 {
#>   margin: 0;
#>   height: 100%;
#> }
#> 
#> .panel-content {
#>   height: 100%;
#>   width: 100%;
#>   padding: var(--card-padding);
#>   grid-area: content;
#> }
#> 
#> .grid_panel .shiny-plot-output > img {
#>   /* The way grid sizing works can throw off the plot sizing in shiny. This
#>      is because the size of the parent grows to its largest child element.
#>      Shiny's plot output uses fixed pixel sizing. Shiny tries to update
#>      these sizes by looking at the parent div's dimensions after a resize.
#>      When the window has been made smaller, this means that the parent
#>      element is spilling outside of its box because the plot image size is
#>      still sitting at a fixed pixel width of the previous size. So when Shiny
#>      queries the size of the parent it thinks that nothing has changed. By
#>      setting a max width and height, we make sure the plot always gets shrunk
#>      to at most the size of the grid element, thus allowing the resize observer
#>      to work properly.
#>   */
#>   max-width: 100%;
#>   max-height: 100%;
#> }
#> 
#> /* Only set card styles if we're not using the bslib card component */
#> .grid_panel:not(.card) {
#>   box-shadow: 0 0 0.5rem rgb(0 0 0 / 35%);
#>   border-radius: 0.5rem;
#> }
#> 
#> 
#> .grid_panel.collapsed  {
#>   height: var(--card-title-height);
#>   overflow: hidden;
#> }
#> 
#> .grid_panel.collapsed .panel-content {
#>   display: none;
#> }
#> 
#> /* Make everything line up nice and cleanly like it should in the middle */
#> .title_panel {
#>   margin: 0;
#>   height: 100%;
#>   display: flex;
#>   align-items: center;
#> }
#> 
#> /* Makes it so tabpanels work in gridpanels */
#> .grid_panel .tabbable { height: 100% }
#> .grid_panel .tabbable > .nav { height: 42px; }
#> .grid_panel .tabbable .tab-content { height: calc(100% - 42px); }
#> .grid_panel .tabbable .tab-pane { height: 100%; }
#> 
#> @media (max-width:600px) {
#> 
#> body {
#>   display: grid;
#>   grid-template-rows: 85px 350px 350px 350px 350px;
#>   grid-template-columns: 1fr;
#>   grid-gap: 1rem;
#>   padding: 1rem;
#>   height: auto;
#> }
#> 
#> #header {
#>   grid-column-start: 1;
#>   grid-column-end: 2;
#>   grid-row-start: 1;
#>   grid-row-end: 2;
#> }
#> #sidebar {
#>   grid-column-start: 1;
#>   grid-column-end: 2;
#>   grid-row-start: 2;
#>   grid-row-end: 3;
#> }
#> #plot_a {
#>   grid-column-start: 1;
#>   grid-column-end: 2;
#>   grid-row-start: 3;
#>   grid-row-end: 4;
#> }
#> #plot_b {
#>   grid-column-start: 1;
#>   grid-column-end: 2;
#>   grid-row-start: 4;
#>   grid-row-end: 5;
#> }
#> #plot_c {
#>   grid-column-start: 1;
#>   grid-column-end: 2;
#>   grid-row-start: 5;
#>   grid-row-end: 6;
#> }
#> 
#> }
#>