Update a layout chunk in a rmd document

Update a layout chunk in a rmd document

update_rmd(
  path_to_rmd,
  new_layout,
  updated_rmd_path = path_to_rmd,
  chunk_tag = "\\.layout \\.grid"
)

Arguments

path_to_rmd
Path to a markdown file with a .layout .grid chunk in it defining a grid layout

new_layout
An updated gridlayout object to replace the current one

updated_rmd_path
Where the rmd with the new layout should be placed. Default is overwriting the previous file

chunk_tag
The identifying labels for the chunk that contains layout. Aka what goes inbetween the curly braces after backticks. This is used as a regular expression so make sure to escape characters properly. For more guidance see regex.

Examples

loc_of_rmd <- system.file("sample_apps/my_app.Rmd", package = "gridlayout")
first_layout <- rmd_to_gridlayout(loc_of_rmd)
first_layout

#> gridlayout object with 2 rows, 2 columns, and gap size: 1rem 
#>      1fr     1fr 
#> 1fr sidebar main
#> 1fr sidebar main

new_layout <- md_to_gridlayout(
  layout_table = "
    |      |120px   |1fr    |1fr    |
    |:-----|:-------|:------|:------|
    |100px |header  |header |header |
    |1fr   |sidebar |plot_a |plot_c |
    |1fr   |sidebar |plot_b |plot_b |"
)

loc_of_new_rmd <- tempfile(fileext = ".rmd")
update_rmd(loc_of_rmd, new_layout, loc_of_new_rmd)

#> Saving rmd with updated layout to /var/folders/d4/cbr0wtz50gb69vd8pqt65w0m0000gn/T//RtmpXIIJfw/file93a33f34d2da.rmd

rmd_to_gridlayout(loc_of_new_rmd)

#> gridlayout object with 3 rows, 3 columns, and gap size: 1rem 
#>        120px   1fr    1fr   
#> 100px header  header header
#> 1fr   sidebar plot_a plot_c
#> 1fr   sidebar plot_b plot_b