Convert layout to css for Shiny

This simply wraps the output of to_css() in a style tag and escapes HTML characters to simplify using in Shiny. Most of the time you'll want to use grid_page or grid_container instead of manually adding this css though.

use_gridlayout_shiny(layout, ...)

Arguments

layout

Either a markdown table representation (see md_to_gridlayout) or a gridlayout object defining the desired layout for your Shiny app.

...

Arguments passed on to to_css

container_key

The unique key used to identify the container to be targeted for the layout. If left blank it will default to applying grid styling to the whole app (aka the body element) for whole page grids. If plain character string is given, then it is assumed to be a gridlayout-key and the targeting is done using an attribute query for data-gridlayout-key. If container contains css selector characters such as a dot, the selector will not be transformed into an id automatically. E.g. container = ".main-content".

Value

Character string of css used to setup grid layout and place elements (referenced by id) into correct locations

See also

Examples

# Only run these examples in interactive R sessions
if (interactive()) {
library(shiny)
my_layout <- "
|      |        |         |
|------|--------|---------|
|2rem  |200px   |1fr      |
|100px |header  |header   |
|500px |sidebar |distPlot |"

# The classic Geyser app with grid layout
shinyApp(
  ui = fluidPage(
    use_gridlayout_shiny(my_layout, "#app-container"),
    div(
      id = "app-container",
      div(
        style = "grid-area: header;",
        h2(id = "app-title", "Old Faithful Geyser Data")
      ),
      div(
        style = "grid-area: sidebar;",
        sliderInput("bins", "Number of bins:",
                    min = 1, max = 50, value = 30
        )
      ),
      div(
        style = "grid-area:distPlot",
        plotOutput("distPlot", height = "100%")
      )
    )
  ),
  server = function(input, output) {
    output$distPlot <- renderPlot({
      x <- faithful[, 2]
      bins <- seq(min(x), max(x), length.out = input$bins + 1)
      hist(x, breaks = bins, col = "darkgray", border = "white")
    })
  }
)

}