Build grid layout obj from markdown table

Build grid layout obj from markdown table

md_to_gridlayout(layout_table)

Arguments

layout_table
Character string with a markdown table. First row and column are reserved for sizing (any valid css sizing works). An optional grid-gap can be provided using the very first cell.

Value

An object of class "grid_layout", which stores the layout as a matrix. This can be passed to other functions such as layout_to_css().

Examples

md_to_gridlayout(
  layout_table = "
    |      |120px   |1fr    |1fr    |
    |:-----|:-------|:------|:------|
    |100px |header  |header |header |
    |1fr   |sidebar |plot_a |plot_c |
    |1fr   |sidebar |plot_b |plot_b |"
)

#> gridlayout object with 3 rows, 3 columns, and gap size: 1rem 
#>        120px   1fr    1fr   
#> 100px header  header header
#> 1fr   sidebar plot_a plot_c
#> 1fr   sidebar plot_b plot_b

# Can specify gap size in upper left cell
md_to_gridlayout(
  layout_table = "
    |25px  |120px  |1fr    |
    |:-----|:------|:------|
    |100px |header |header |
    |1fr   |plot   |table  |
    |1fr   |footer |footer |"
)

#> gridlayout object with 3 rows, 2 columns, and gap size: 25px 
#>        120px  1fr   
#> 100px header header
#> 1fr   plot   table 
#> 1fr   footer footer

# Don't need to follow full md table with
# header row if not desired
md_to_gridlayout(
  layout_table = "
    |25px  |120px  |1fr    |
    |100px |header |header |
    |1fr   |plot   |table  |
    |1fr   |footer |footer |"
)

#> gridlayout object with 3 rows, 2 columns, and gap size: 25px 
#>        120px  1fr   
#> 100px header header
#> 1fr   plot   table 
#> 1fr   footer footer

# Can omit sizing as well if desired
md_to_gridlayout(
  layout_table = "
    |header |header |
    |plot   |table  |
    |footer |footer |"
)

#> gridlayout object with 3 rows, 2 columns, and gap size: 1rem 
#>      1fr    1fr   
#> 1fr header header
#> 1fr plot   table 
#> 1fr footer footer