Construct a gridlayout object from basic parts

Construct a gridlayout object from basic parts

new_gridlayout(
  layout_mat,
  col_sizes,
  row_sizes,
  gap,
  element_list,
  warn_about_overap = FALSE
)

Arguments

layout_mat
A character matrix where each entry controls what element id will take up that given position in the grid. For instance a header element would take up every cell in the first row.

col_sizes
A character vector of valid css sizes for the width of each column in your grid as given by layout_mat. If a single value is passed, it will be repeated for all columns.

row_sizes
Same as col_sizes, but for row heights.

gap
Valid css sizing for gap to be left between each element in your grid. Defaults to "1rem".

element_list
List of elements with the id, start_row, end_row, start_col, and end_col format.

warn_about_overap
Should overlapping elements in the grid be flagged? Only used in element_list argument is use to define layout

Value

Object of class "gridlayout"

Examples

my_layout <- new_gridlayout(
  layout_mat = matrix(c(
    "header", "header",
    "plota",  "plotb"
  ), ncol = 2, byrow = TRUE),
  col_sizes = c("1fr", "2fr"),
  row_sizes = c("100px", "1fr"),
  gap = "2rem"
)