Convert grid layout to css string

Convert grid layout to css string

to_css(layout, container_key, is_card_styled = "grid_panel")

Arguments

layout

Object of class "gridlayout".

container_key

The unique key used to identify the container to be targeted for the layout. If left blank it will default to applying grid styling to the whole app (aka the body element) for whole page grids. If plain character string is given, then it is assumed to be a gridlayout-key and the targeting is done using an attribute query for data-gridlayout-key. If container contains css selector characters such as a dot, the selector will not be transformed into an id automatically. E.g. container = ".main-content".

is_card_styled

Should each section of the grid be given a card style to make it stand out? Options are "grid_panel", where only elements with "grid_panel" class will get card styling, "all" where all children of the grid container will get card styling (useful for RMarkdown or other situations where you don't control child rendering) or "none" for no card styling.

selector_prefix

CSS prefix used to target grid elements. This will change if you're integrating grid with a system that you don't want to use ids (the "#" prefix) with because they are not available or are used for other reasons.

Value

Character string of css used to setup grid layout and place elements (referenced by id) into correct locations

Examples

grid_obj <- md_to_gridlayout(
  layout_table = "
    |      |120px   |1fr    |1fr    |
    |:-----|:-------|:------|:------|
    |100px |header  |header |header |
    |1fr   |sidebar |plot_a |plot_c |
    |1fr   |sidebar |plot_b |plot_b |"
)

cat(to_css(grid_obj))
#> 
#> body {
#>   display:grid;
#>   grid-template-rows:100px 1fr 1fr;
#>   grid-template-columns:120px 1fr 1fr;
#>   grid-template-areas:
#>     "header  header header"
#>     "sidebar plot_a plot_c"
#>     "sidebar plot_b plot_b";
#>   grid-gap:1rem;
#>   padding:1rem;
#>   height:100vh;
#> }
#> 
#> 
#> 
#> @media (max-width:500px) {
#>   body {
#>     display:grid;
#>     grid-template-rows:85px 350px 350px 350px 350px;
#>     grid-template-columns:1fr;
#>     grid-template-areas:
#>       "header "
#>       "sidebar"
#>       "plot_a "
#>       "plot_b "
#>       "plot_c ";
#>     grid-gap:1rem;
#>     padding:1rem;
#>     height:auto;
#>   }
#> }
#>