Convert layout to css for Shiny

This simply wraps the output of to_css() in a style tag and escapes HTML characters to simplify using in Shiny. Most of the time you'll want to use grid_page or grid_container instead of manually adding this css though.

use_gridlayout_shiny(layout, ...)

Arguments

layout
Either a markdown table representation (see md_to_gridlayout) or a gridlayout object defining the desired layout for your Shiny app.

...
Arguments passed on to to_css

container

Id of the element for grid to be placed in. Default value to apply grid styling to the whole app (aka the body element) for whole page grids. If an simple name is passed it is assumed that it is an id: i.e. the css id selector # is apended to the name. If container contains css selector characters such as a dot, the selector will not be transformed into an id automatically. E.g. container = ".main-content".

is_card_styled

Should each section of the grid be given a card style to make it stand out? Options are "grid_panel", where only elements with "grid_panel" class will get card styling, "all" where all children of the grid container will get card styling (useful for RMarkdown or other situations where you don't control child rendering) or "none" for no card styling.

element_styles

A list of named property-value pairs ( E.g. element_styles = c("background" = "blue").) for additional styles to be added to each first child element of the grid container. E.g. under the css selector of grid_container > *. If is_card_styled = "all" then this has the same purpose as as additional_card_styles

selector_prefix

CSS prefix used to target grid elements. This will change if you're integrating grid with a system that you don't want to use ids (the "#" prefix) with because they are not available or are used for other reasons.

Value

Character string of css used to setup grid layout and place elements (referenced by id) into correct locations

See also

Examples

# Only run these examples in interactive R sessions
if (interactive()) {

my_layout <- "
|      |        |         |
|------|--------|---------|
|2rem  |200px   |1fr      |
|100px |header  |header   |
|1fr   |sidebar |distPlot |"

# The classic Geyser app with grid layout
shinyApp(
  ui = fluidPage(
    use_gridlayout_shiny(my_layout, "app-container"),
    div(
      id = "app-container",
      div(id = "header",
          h2(id = "app-title", "Old Faithful Geyser Data")
      ),
      div(id = "sidebar",
          sliderInput("bins","Number of bins:",
                      min = 1, max = 50, value = 30)
      ),
      plotOutput("distPlot")
    )
  ),
  server = function(input, output) {
    output$distPlot <- renderPlot({
      x    <- faithful[, 2]
      bins <- seq(min(x), max(x), length.out = input$bins + 1)
      hist(x, breaks = bins, col = 'darkgray', border = 'white')
    })
  }
)
}